backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
python-dotenv==1.0.1
supabase==2.9.1
PyMuPDF==1.24.10
python-docx==1.1.2
openpyxl==3.1.5
Pillow==10.4.0
pytesseract==0.3.13
pyjwt==2.9.0
httpx==0.27.2
